RequirementMachine: Move term to type conversion to a new file InterfaceType.cpp
